move navigation to own component & wrap with Suspense
